fix(adapters): gate testcontainers-go test fixture behind nodockerfixture build tag (#929)
`probeDockerAvailable`, `startGrypeOfflineDBContainer`, `NewGrypeAdapterFixedDB`, and `NewGrypeAdapterFixedDBWithMatchers` (`adapters/v1/grype.go`) are a Docker-container-backed test fixture used by `grype_test.go` and `scan_test.go` to test GrypeAdapter against a real, pinned vulnerability DB container. Living in an ordinary `.go` file meant `testcontainers-go` and its dependency tree (`github.com/docker/docker` and friends) were compiled into the production `kubevuln` binary for zero runtime benefit, causing false-positive vulnerability findings (GO-2026-4887 and GO-2026-4883) in CI binary security scans.

Move the Docker test fixture functions into `adapters/v1/grype_docker_fixture.go` gated with `//go:build !nodockerfixture`. Pass `-tags nodockerfixture` in `Makefile`'s build target and `build/Dockerfile` / `build/Dockerfile.debug` so the shipped production binary excludes `testcontainers-go` and `github.com/docker/docker/client`. Add regression test `TestProductionBuildExcludesDockerFixture` in `cmd/http/main_test.go` asserting that `testcontainers-go` and `docker/docker/client` are excluded from production build dependencies.

Closes #929
